% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_cde_legacy_url}
\alias{build_cde_legacy_url}
\title{Build CDE URL for legacy enrollment files (2009-2018)}
\usage{
build_cde_legacy_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
URL string
}
\description{
Build CDE URL for legacy enrollment files (2009-2018)
}
\keyword{internal}
