% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{clean_assessment_column_names}
\alias{clean_assessment_column_names}
\title{Clean assessment column names}
\usage{
clean_assessment_column_names(x)
}
\arguments{
\item{x}{Character vector of column names}
}
\value{
Cleaned column names
}
\description{
Standardizes column names from CMAS Excel files.
}
\keyword{internal}
