% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_school_id}
\alias{create_school_id}
\title{Create combined school ID}
\usage{
create_school_id(district_code, school_code)
}
\arguments{
\item{district_code}{District code}

\item{school_code}{School code}
}
\value{
8-character ID (district + school)
}
\description{
Combines district and school codes into a single unique identifier.
}
\keyword{internal}
