% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_state_aggregates_from_districts}
\alias{create_state_aggregates_from_districts}
\title{Create state-level aggregates from district data}
\usage{
create_state_aggregates_from_districts(district_tidy)
}
\arguments{
\item{district_tidy}{Long format district-level data}
}
\value{
Data frame of state-level tidy rows
}
\description{
Create state-level aggregates from district data
}
\keyword{internal}
