% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_file}
\alias{download_assessment_file}
\title{Download a single assessment file}
\usage{
download_assessment_file(end_year, subject)
}
\arguments{
\item{end_year}{School year end}

\item{subject}{Subject code (ela, math, science, csla)}
}
\value{
Data frame or NULL if download fails
}
\description{
Download a single assessment file
}
\keyword{internal}
