% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_school_year}
\alias{format_school_year}
\title{Format end year as school year string}
\usage{
format_school_year(end_year)
}
\arguments{
\item{end_year}{End year (e.g., 2024)}
}
\value{
School year string (e.g., "2023-24")
}
\description{
Format end year as school year string
}
\keyword{internal}
