% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from CDE}
\usage{
get_raw_directory()
}
\value{
A list with two data frames: schools and districts
}
\description{
Downloads the raw school directory Excel files from the Colorado
Department of Education CEDAR library. Downloads both school addresses
and district addresses.
}
\keyword{internal}
