% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_year}
\alias{parse_school_year}
\title{Convert school year string to end year}
\usage{
parse_school_year(year_string)
}
\arguments{
\item{year_string}{School year string (e.g., "2023-24")}
}
\value{
Integer end year (e.g., 2024)
}
\description{
Converts various school year formats (2023-24, 2023-2024, etc.) to end year.
}
\keyword{internal}
