% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{pivot_to_long}
\alias{pivot_to_long}
\title{Pivot wide data to long format}
\usage{
pivot_to_long(df, subgroups)
}
\arguments{
\item{df}{Wide data frame}

\item{subgroups}{Character vector of subgroup column names}
}
\value{
Long format data frame
}
\description{
Pivot wide data to long format
}
\keyword{internal}
