% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_district_addresses}
\alias{process_district_addresses}
\title{Process district addresses data}
\usage{
process_district_addresses(data)
}
\arguments{
\item{data}{Raw district addresses data frame}
}
\value{
Processed data frame
}
\description{
Process district addresses data
}
\keyword{internal}
