% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr}
\alias{process_enr}
\title{Process raw enrollment data to wide format}
\usage{
process_enr(df, end_year)
}
\arguments{
\item{df}{Raw data frame from download_cde_excel}

\item{end_year}{School year end}
}
\value{
Data frame in wide format with standardized columns
}
\description{
Transforms the raw Excel data into a standardized wide format with proper
column names and data types. Colorado's data has a 2-row header structure
}
\keyword{internal}
