% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_school_addresses}
\alias{process_school_addresses}
\title{Process school addresses data}
\usage{
process_school_addresses(data)
}
\arguments{
\item{data}{Raw school addresses data frame}
}
\value{
Processed data frame
}
\description{
Process school addresses data
}
\keyword{internal}
