% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{read_assessment_excel}
\alias{read_assessment_excel}
\title{Read assessment data from Excel file}
\usage{
read_assessment_excel(filepath, end_year, subject)
}
\arguments{
\item{filepath}{Path to downloaded Excel file}

\item{end_year}{School year end}

\item{subject}{Subject code}
}
\value{
Data frame with assessment data
}
\description{
Parses the CMAS assessment Excel file.
}
\keyword{internal}
