% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_assessment_grade}
\alias{standardize_assessment_grade}
\title{Standardize grade levels}
\usage{
standardize_assessment_grade(x)
}
\arguments{
\item{x}{Vector of grade values}
}
\value{
Standardized grade levels
}
\description{
Standardize grade levels
}
\keyword{internal}
