% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_school_code}
\alias{standardize_school_code}
\title{Standardize Colorado school code}
\usage{
standardize_school_code(code)
}
\arguments{
\item{code}{School code (numeric or character)}
}
\value{
Character string of 4-digit code
}
\description{
Colorado school codes are typically 4 digits.
}
\keyword{internal}
